fix: mask-based skip, float64 precision, load_network_params mask, OBD diagonal, omega warning
Bugs fixed:
- neuron_sparsifier: missing hidden_activation setup in main() (tanh/sigmoid experiments used wrong activation)
- All Python sparsifiers: skip condition uses mask==0 instead of W==0
- prune_ext.cpp: candidate filter uses mask!=0 instead of W!=0
- mlp.py: load_network_params reconstructs mask=(W!=0) instead of ones_like
- Float64 precision: C++ extension (forward.hpp, prune_ext.cpp) upgraded from float32 to double; make_omega and C++ call-sites now pass float64 (JAX x64 was already enabled)
- lazarevich_sparsifier: warning when omega_samples > available calibration images; omega cast to float64
- obd_sparsifier: _diag_hessian replaced jax.hessian (O(N²) memory) with N forward-JVP calls, peak memory O(N)

+++ prune_ext/activations.hpp
@@ -5,15 +5,15 @@
 #include <string>
 #include <unordered_map>
 
-using ActivFn = float (*)(float);
+using ActivFn = double (*)(double);
 
 // Returns the activation function pointer for the given name.
 inline ActivFn get_activation(const std::string& name) {
     static const std::unordered_map<std::string, ActivFn> reg = {
-        {"relu",    [](float x) -> float { return x > 0.f ? x : 0.f; }},
-        {"tanh",    [](float x) -> float { return std::tanh(x); }},
-        {"sigmoid", [](float x) -> float { return 1.f / (1.f + std::exp(-x)); }},
-        {"linear",  [](float x) -> float { return x; }},
+        {"relu",    [](double x) -> double { return x > 0.0 ? x : 0.0; }},
+        {"tanh",    [](double x) -> double { return std::tanh(x); }},
+        {"sigmoid", [](double x) -> double { return 1.0 / (1.0 + std::exp(-x)); }},
+        {"linear",  [](double x) -> double { return x; }},
         // add new activations here
     };
     auto it = reg.find(name);
